mkknlimg: Fix for bcm2708-pinctrl removal and unxz breakage
In the latest configurations, bcm2708-pinctrl has been replaced
by bcm2835-gpio. Modify the script to also recognise the latter.

Independently, a patch to fix unxz support broke (all?) other
compression methods. Solve the problem of multiple matches by
adding a parameter for each compression method indicating which
match to use. Set all to '0' (first) except unxz, which gets
'-1' (last) to match the 'tail -n 1' command it replaces.
[ Would '1' (second) be better? ]

In the event that the script doesn't recognise a valid bcm2708
kernel, it will copy the input file to the output without
modification AND return an error code.