Remove the knlinfo and mkknlimg tools
The canonical home of the mkknlimg and knlinfo tools is now the Linux
tree in the scripts directory. Different branches may contain different
versions making external copies such as these likely to be inconsistent
with at least one of them, so delete the the copies.

See: https://github.com/raspberrypi/tools/issues/58