armstub7: Disable data cache and MMU
The previous code was enabling both I and D-caches, but keeping the
bit M as the reset value, which is architecturally undetermined. In
practice, in both Cortex-A7 and Cortex-A53, the reset value is 0, so
the D-cache was disabled (but the I-cache was enabled).

For the 32-bit Arm architecture boot, the Linux kernel expects the data
cache and MMU to be disabled when the CPU jumps into it, the I-cache
doesn't matter, as specified in the file `Documentation/arm/Booting` of
the Linux kernel tree under `Calling the kernel image`.

This patch disables the D-cache and MMU for correctness while keeping
the I-cache enabled to save a few cycles during the execution of the
stub.

Signed-off-by: Antonio Niño Díaz <antonio_nd@outlook.com>