armstub7: Fix reset NSACR values
In Cortex-A7, some reserved fields of NSACR were set to 1. According to
the Technical Reference Manual, the reserved fields of this register are
RAZ/WI (Read-As-Zero, Writes Ignored).

In Cortex-A53, all reserved fields are RES0. Software should respect
this for future compatibility.

The bits NS_SMP and NS_L2ERR of Cortex-A7 need to be set to 1 in order
for the Non-secure world to be able to modify the value of the SMP and
L2 AXI asynchronous error bits. However, in Cortex-A53 they are RES0.
This means that both CPUs can't share the same reset value.

Signed-off-by: Antonio Niño Díaz <antonio_nd@outlook.com>